QFJ-802 Change mvn dependancies from messages -> core  to core -> messages Move classes required to compile FIX-generated code from core to messages-fix Replace FIX code generation + compilation duplicated in both "messages" and "fix" module with one single in "messages"
