+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/org/quickfixj/CharsetSupport.java
