+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/quickfix/DoubleField.java
