+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/quickfix/Field.java
@@ -19,10 +19,9 @@
 
 package quickfix;
 
-import org.quickfixj.CharsetSupport;
-
 import java.io.Serializable;
-import java.nio.charset.Charset;
+
+import org.quickfixj.CharsetSupport;
 
 /**
  * Base class for FIX message fields. This class should be
