+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/quickfix/MessageComponent.java
