+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/quickfix/MessageUtils.java
@@ -29,7 +29,6 @@ import org.quickfixj.QFJException;
 import quickfix.Message.Header;
 import quickfix.field.ApplVerID;
 import quickfix.field.BeginString;
-import quickfix.field.DefaultApplVerID;
 import quickfix.field.MsgType;
 import quickfix.field.SenderCompID;
 import quickfix.field.SenderLocationID;
@@ -114,74 +113,7 @@ public class MessageUtils {
         return message;
     }
 
-    /**
-     * NOTE: This method is intended for internal use.
-     *
-     * @param session the Session that will process the message
-     * @param messageString
-     * @return the parsed message
-     * @throws InvalidMessage
-     */
-    public static Message parse(Session session, String messageString) throws InvalidMessage {
-        final String beginString = getStringField(messageString, BeginString.FIELD);
-        final String msgType = getMessageType(messageString);
-
-        ApplVerID applVerID;
-
-        if (FixVersions.BEGINSTRING_FIXT11.equals(beginString)) {
-            applVerID = getApplVerID(session, messageString);
-        } else {
-            applVerID = toApplVerID(beginString);
-        }
-
-        final MessageFactory messageFactory = session.getMessageFactory();
-
-        final DataDictionaryProvider ddProvider = session.getDataDictionaryProvider();
-        final DataDictionary sessionDataDictionary = ddProvider == null ? null : ddProvider
-                .getSessionDataDictionary(beginString);
-        final DataDictionary applicationDataDictionary = ddProvider == null ? null : ddProvider
-                .getApplicationDataDictionary(applVerID);
-
-        final quickfix.Message message = messageFactory.create(beginString, msgType);
-        final DataDictionary payloadDictionary = MessageUtils.isAdminMessage(msgType)
-                ? sessionDataDictionary
-                : applicationDataDictionary;
-
-        message.parse(messageString, sessionDataDictionary, payloadDictionary,
-                payloadDictionary != null);
-
-        return message;
-    }
-
-    private static ApplVerID getApplVerID(Session session, String messageString)
-            throws InvalidMessage {
-        ApplVerID applVerID = null;
-
-        final String applVerIdString = getStringField(messageString, ApplVerID.FIELD);
-        if (applVerIdString != null) {
-            applVerID = new ApplVerID(applVerIdString);
-        }
-
-        if (applVerID == null) {
-            applVerID = session.getTargetDefaultApplicationVersionID();
-        }
-
-        if (applVerID == null && isLogon(messageString)) {
-            final String defaultApplVerIdString = getStringField(messageString,
-                    DefaultApplVerID.FIELD);
-            if (defaultApplVerIdString != null) {
-                applVerID = new ApplVerID(defaultApplVerIdString);
-            }
-        }
-
-        if (applVerID == null) {
-            throw new InvalidMessage("Can't determine ApplVerID for message");
-        }
-
-        return applVerID;
-    }
-
-    public static boolean isAdminMessage(String msgType) {
+        public static boolean isAdminMessage(String msgType) {
         return msgType.length() == 1 && "0A12345".contains(msgType);
     }
