+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/quickfix/StringField.java
