+++ quickfixj-messages/quickfixj-messages-fix/src/main/java/quickfix/field/converter/UtcDateOnlyConverter.java
