+++ quickfixj-messages/quickfixj-messages-fix50sp1/pom.xml
@@ -19,6 +19,14 @@
 		<fix.name>fix50sp1</fix.name>
 	</properties>
 
+	<dependencies>
+		<dependency>
+			<groupId>org.quickfixj</groupId>
+			<artifactId>quickfixj-messages-fix</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+	</dependencies>
+	
 	<build>
 		<plugins>
 			<plugin>
